+++ go.mod
@@ -1,11 +1,3 @@
 module github.com/lionslon/yapmetrics
 
 go 1.20
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/stretchr/objx v0.5.0 // indirect
-	github.com/stretchr/testify v1.8.4 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
